Add configuration file to save selected mode and version
The UIA version and selected modes are saved in the App.config.
If the mode and version is set at application start, then both is set.
Also when switching the mode it is saved in the config.
Additionally, when the UIA version wants to be changed then the new menu entry is selected